The problem was detected and the process for finding the solution is started.
