+++ src/index.ts
@@ -1,7 +1,7 @@
-import GeoJSONSource from "./GeoJSONSource";
-import Map from "./Map";
-import Layer from "./Layer";
-import Marker from "./Marker";
+import GeoJSONSource from "./components/GeoJSONSource";
+import Map from "./components/Map";
+import Layer from "./components/Layer";
+import Marker from "./components/Marker";
 import ViewPort from "./ViewPort";
 
 export { Map, Layer, Marker, ViewPort, GeoJSONSource };
